event cache: get rid of the `RoomNotFound` error
By attaching the `RoomEventCache` public getter to the `Room` itself, we can make it so that it's impossible to get an error when getting the room
cache for a room. The only downside is that, when receiving room updates, we now need to fetch the room in the client, and log an error if it's
missing. It really shouldn't be missing at any time, so as when we're calling event handlers we're logging out an `error!` when that happens.